[redfish_test] refactoring + in-memory complete configuration tree
%%% Code status
 |
 ----------> In development
 |
 ----------> Compiling

%%% Level of test
 |
 ----------> Hand-tested

%%% Comments
 |
 ------
      |
      v
The existing test file was completely refactored so as to make it:
    + compilable;
    + readable;
    + coherent with the plugin;
    + suitable for the implementation of a complete test suite.

On top of that, a complete configuration file was designed.
The latter is associated with the "bladed 1U server" mock proposed by
DMTF.
It uses the full range of features exposed by the Redfish plugin for
@p`collectd`.
An in-memory copy of this configuration file was built/implemented in
@f`redfish_test` so as to make the implementation of the test suite
easy/clear/fast.